Remove Nth node from end while only 1 traversal
-	At beginning of program, jump n nodes. This ensures that removing the
node at head.next will be the nth node from the end (right).
-	Change recursion to iteration to try to improve memory performance.
